+++ CHANGELOG.md
@@ -58,3 +58,8 @@
 - Found: step_1_data_preparation/step1.5/run_validate_data.sh
 - Found: step_1_data_preparation/step1.5/split_train_val_test.py
 - Found: step_1_data_preparation/step1.5/validate_final_data.py
+
+## [2025-11-19 14:56:38] Auto Commit
+
+- Found: step_2_model_architecture/LMOD.sourceme
+- Found: step_2_model_architecture/step2.1/